Refactors the sort dropdown form into its own component. Makes everything much neater
